RequestHandlerMixin: handle toggling "sampled".
When the "sampled" flag is changed on a span, then any future children
should be sampled as well.  We were not updating the propagated headers
if sampled was changed after `prepare' had finished.  I added a new
property of the RequestHandler named `request_is_traced' that can be
modified to enable or disable sampling of the handler's span.